Using int32_t for rank for larger jobs with 32k processes or more
git-svn-id: https://scm.mvapich.cse.ohio-state.edu/svn/mpi/mvapich2/trunk@2892 09bc9535-d30e-0410-b1f7-d46b20a4725c

+++ src/mpid/ch3/include/mpidpre.h
@@ -65,7 +65,7 @@ typedef unsigned long MPID_Seqnum_t;
 typedef struct MPIDI_Message_match
 {
     int32_t tag;
-    int16_t rank;
+    int32_t rank;
     int16_t context_id;
 }
 MPIDI_Message_match;
